Move workout fetching into server component
